% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.mvoprobit}
\alias{fitted.mvoprobit}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{mvoprobit}(object, ..., newdata = NULL)
}
\arguments{
\item{object}{object of class 'mvoprobit'.}

\item{...}{further arguments (currently ignored).}

\item{newdata}{an optional data frame in which to look for variables 
with which to predict. If omitted, the original data frame used. 
This data frame should contain values of dependent variables even if they 
are not actually needed for prediction (simply assign them with 0 values).}
}
\value{
Returns a data frame. Its columns which names coincide with the
names of the ordered equations provide an index of the most probable
category. Columns which names coincide with the
names of the continuous equations provide uncinditional expectations of
the dependent variables in available regimes.
}
\description{
Extracts fitted values from 'mvoprobit' object
}
