% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula.mnprobit}
\alias{formula.mnprobit}
\title{Formulas of mnprobit model.}
\usage{
\method{formula}{mnprobit}(x, ..., type = "formula")
}
\arguments{
\item{x}{object of class 'mnprobit'.}

\item{...}{further arguments (currently ignored).}

\item{type}{character; 
if \code{type = "formula"} or \code{type = 1} then function returns 
a formulas of multinomial equation. 
If \code{type = "formula2"} or \code{type = 2} then function returns 
a formula of continuous equation.}
}
\value{
Returns a formula.
}
\description{
Provides formulas associated with the 
object of class 'mnprobit'.
}
