% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula.msel}
\alias{formula.msel}
\title{Formulas of msel model.}
\usage{
\method{formula}{msel}(x, ..., type = "formula", eq = NULL)
}
\arguments{
\item{x}{object of class 'msel'.}

\item{...}{further arguments (currently ignored).}

\item{type}{character; 
if \code{type = "formula"} or \code{type = 1} then function returns formulas 
of the ordinal equations. 
If \code{type = "formula2"} or \code{type = 2} then function returns formulas 
of the continuous equations.
If \code{type = "formula3"} or \code{type = 3} then function returns formula 
of the multinomial equation.}

\item{eq}{positive integer representing the index of the equation which
formula should be returned. If \code{NULL} (default) then formulas for each
equation will be returned as a list which \code{i}-th element associated 
with \code{i}-th equation.}
}
\value{
Returns a formula or a list of formulas depending on \code{eq} value.
}
\description{
Provides formulas associated with the 
object of class 'msel'.
}
