% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula.mvoprobit}
\alias{formula.mvoprobit}
\title{Formulas of mvoprobit model.}
\usage{
\method{formula}{mvoprobit}(x, ..., type = "formula", eq = NULL)
}
\arguments{
\item{x}{object of class 'mvoprobit'.}

\item{...}{further arguments (currently ignored).}

\item{type}{character; 
if \code{type = "formula"} or \code{type = 1} then function returns formulas 
of ordered equations. 
If \code{type = "formula2"} or \code{type = 2} then function returns formulas 
of continuous equations.}

\item{eq}{positive integer representing the index of the equation which
formula should be returned. If \code{NULL} (default) then formulas for each
equation will be returned as a list which \code{i}-th element associated 
with \code{i}-th equation.}
}
\value{
Returns a formula or a list of formulas depending on \code{eq} value.
}
\description{
Provides formulas associated with the 
object of class 'mvoprobit'.
}
