% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lnL_mvoprobit}
\alias{lnL_mvoprobit}
\title{Log-likelihood Function of Multivariate Ordered Probit Model}
\usage{
lnL_mvoprobit(par, control_lnL, out_type = "val", n_sim = 1000L, n_cores = 1L)
}
\arguments{
\item{par}{vector of parameters.}

\item{control_lnL}{list with some additional parameters.}

\item{out_type}{string represeint the output type of the function.}

\item{n_sim}{the number of random draws for multivariate 
normal probabilities.}

\item{n_cores}{the number of cores to be used.}
}
\description{
Calculates log-likelihood function of multivariate ordered
probit model.
}
