% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.msel}
\alias{logLik.msel}
\title{Extract Log-Likelihood from a Fit of the msel Function.}
\usage{
\method{logLik}{msel}(object, ...)
}
\arguments{
\item{object}{object of class "msel"}

\item{...}{further arguments (currently ignored)}
}
\value{
Returns an object of class 'logLik'.
}
\description{
Extract Log-Likelihood from a model fit
of the \code{\link[switchSelection]{msel}} function.
}
\details{
If \code{estimator == "2step"} in 
\code{\link[switchSelection]{msel}} then function may return
\code{NA} value since the two-step estimator of the covariance matrix may not
be positively defined.
}
