% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.mvoprobit}
\alias{logLik.mvoprobit}
\title{Extract Log-Likelihood from a Fit of the mvoprobit Function.}
\usage{
\method{logLik}{mvoprobit}(object, ...)
}
\arguments{
\item{object}{object of class "mvoprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
Returns an object of class 'logLik'.
}
\description{
Extract Log-Likelihood from a model fit
of the \code{\link[switchSelection]{mvoprobit}} function.
}
\details{
If \code{estimator == "2step"} in 
\code{\link[switchSelection]{mvoprobit}} then function may return
\code{NA} value since two-step estimator of covariance matrix may be
not positively defined.
}
