% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{nobs.mnprobit}
\alias{nobs.mnprobit}
\title{Extract the Number of Observations from a Fit of the mnprobit Function.}
\usage{
\method{nobs}{mnprobit}(object, ...)
}
\arguments{
\item{object}{object of class "mnprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
A single positive integer number.
}
\description{
Extract the number of observations from a model fit
of the \code{\link[switchSelection]{mnprobit}} function.
}
\details{
Unobservable values of continuous equations are included into
the number of observations.
}
