% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoprobit.R
\name{nobs.mvoprobit}
\alias{nobs.mvoprobit}
\title{Extract the Number of Observations from a Fit of the mvoprobit Function.}
\usage{
\method{nobs}{mvoprobit}(object, ...)
}
\arguments{
\item{object}{object of class "mvoprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
A single positive integer number.
}
\description{
Extract the number of observations from a model fit
of the \code{\link[switchSelection]{mvoprobit}} function.
}
\details{
Unobservable values of continuous equations are included into
the number of observations.
}
