% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{print.lrtest}
\alias{print.lrtest}
\title{Print Method for Likelihood Ratio Test}
\usage{
\method{print}{lrtest}(x, ...)
}
\arguments{
\item{x}{object of class "lrtest".}

\item{...}{further arguments (currently ignored).}
}
\value{
The function returns input argument \code{x}.
}
\description{
Prints summary for an object of class 'lrtest'.
}
