% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{print.lrtest_msel}
\alias{print.lrtest_msel}
\title{Print Method for Likelihood Ratio Test}
\usage{
\method{print}{lrtest_msel}(x, ...)
}
\arguments{
\item{x}{object of class "summary.lrtest_msel".}

\item{...}{further arguments (currently ignored).}
}
\value{
The function returns the input argument \code{x}.
}
\description{
Prints summary for an object of class 'lrtest_msel'.
}
