% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.msel}
\alias{print.msel}
\title{Print for an Object of Class msel}
\usage{
\method{print}{msel}(x, ...)
}
\arguments{
\item{x}{object of class 'msel'}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns \code{NULL}.
}
\description{
Prints information on the object of class 'msel'.
}
