% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoprobit.R
\name{print.mvoprobit}
\alias{print.mvoprobit}
\title{Print for an Object of Class mvoprobit}
\usage{
\method{print}{mvoprobit}(x, ...)
}
\arguments{
\item{x}{object of class "mvoprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns input argument \code{x}.
}
\description{
Prints summary for an object of class 'mvoprobit'.
}
