% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.struct_msel}
\alias{print.struct_msel}
\title{Print for an Object of Class struct_msel}
\usage{
\method{print}{struct_msel}(x, ...)
}
\arguments{
\item{x}{object of class 'struct_msel'}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns \code{NULL}.
}
\description{
Prints information on the object of class 'struct_msel'.
}
