% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{print.summary.delta_method}
\alias{print.summary.delta_method}
\title{Print summary for an Object of Class delta_method}
\usage{
\method{print}{summary.delta_method}(x, ...)
}
\arguments{
\item{x}{object of class 'delta_method'}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns input argument \code{x}.
}
\description{
Prints summary for an object of class 'delta_method'.
}
