% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{print.summary.lrtest}
\alias{print.summary.lrtest}
\title{Print Summary Method for Likelihood Ratio Test}
\usage{
\method{print}{summary.lrtest}(x, ...)
}
\arguments{
\item{x}{object of class "lrtest"}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns input argument \code{x} changing
it's class to \code{lrtest}.
}
\description{
Prints summary for an object of class 'lrtest'.
}
