% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{print.summary.lrtest_msel}
\alias{print.summary.lrtest_msel}
\title{Print Summary Method for Likelihood Ratio Test}
\usage{
\method{print}{summary.lrtest_msel}(x, ...)
}
\arguments{
\item{x}{object of class "lrtest_msel"}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns input argument \code{x} changing
its class to \code{lrtest_msel}.
}
\description{
Prints summary for an object of class 'lrtest_msel'.
}
