% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.mnprobit}
\alias{print.summary.mnprobit}
\title{Print summary for an Object of Class mnprobit}
\usage{
\method{print}{summary.mnprobit}(x, ...)
}
\arguments{
\item{x}{object of class 'mnprobit'}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns \code{x}.
}
\description{
Prints summary for an object of class 'mnprobit'.
}
