% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.msel}
\alias{print.summary.msel}
\title{Print summary for an Object of Class msel}
\usage{
\method{print}{summary.msel}(x, ...)
}
\arguments{
\item{x}{object of class 'msel'}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns \code{x}.
}
\description{
Prints summary for an object of class 'msel'.
}
