% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{print.summary.test_msel}
\alias{print.summary.test_msel}
\title{Print summary for an Object of Class test_msel}
\usage{
\method{print}{summary.test_msel}(x, ..., is_legend = TRUE)
}
\arguments{
\item{x}{object of class 'test_msel'}

\item{...}{further arguments (currently ignored)}

\item{is_legend}{a logical; if \code{TRUE} then additional information
is shown.}
}
\value{
The function returns input argument \code{x}.
}
\description{
Prints summary for an object of class 'test_msel'.
}
