% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma.mnprobit}
\alias{sigma.mnprobit}
\title{Extract Residual Standard Deviation 'Sigma'}
\usage{
\method{sigma}{mnprobit}(object, use.fallback = TRUE, ..., regime = NULL)
}
\arguments{
\item{object}{object of class "mnprobit".}

\item{use.fallback}{logical, passed to \code{nobs} (currently ignored).}

\item{...}{further arguments (currently ignored).}

\item{regime}{regime of continuous equation}
}
\value{
Returns the estimates of standard deviations 
of \eqn{\varepsilon_{i}}.
If \code{regime = r} then estimate of 
\eqn{\sqrt{Var(\varepsilon_{ri})}} is returned.
}
\description{
Extract standard deviations of random errors of continuous
equation of \code{\link[switchSelection]{mnprobit}} function.
}
\details{
Available only if \code{estimator = "ml"}.
}
