% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{starsVector}
\alias{starsVector}
\title{Stars for p-values}
\usage{
starsVector(p_value)
}
\arguments{
\item{p_value}{vector of values between 0 and 1 representing p-values.}
}
\value{
The function returns a string vector of stars assigned according
to the rules described in 'Details' section.
}
\description{
This function assigns stars 
(associated with different significance levels) to p-values.
}
\details{
Three stars are assigned to p-values not greater than 0.01.
Two stars are assigned to p-values greater than 0.01 and not greater 
than 0.05.
One star is assigned to p-values greater than 0.05 and not greater than 0.1.
}
\examples{
p_value <- c(0.002, 0.2, 0.03, 0.08)
starsVector(p_value)
}
