% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{summary.delta_method}
\alias{summary.delta_method}
\title{Summary for an Object of Class delta_method}
\usage{
\method{summary}{delta_method}(object, ...)
}
\arguments{
\item{object}{object of class 'delta_method'}

\item{...}{further arguments (currently ignored)}
}
\value{
Returns an object of class 'summary.delta_method'.
}
\description{
Provides summary for an object of class 'delta_method'.
}
