% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{summary.lrtest}
\alias{summary.lrtest}
\title{Summary Method for Likelihood Ratio Test}
\usage{
\method{summary}{lrtest}(object, ...)
}
\arguments{
\item{object}{object of class "lrtest"}

\item{...}{further arguments (currently ignored)}
}
\value{
Returns an object of class 'summary.lrtest'.
}
\description{
Provides summary for an object of class 'lrtest'.
}
\details{
This function just changes the class of the 'lrtest'
object to 'summary.lrtest'.
}
