% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{summary.lrtest_msel}
\alias{summary.lrtest_msel}
\title{Summary Method for Likelihood Ratio Test}
\usage{
\method{summary}{lrtest_msel}(object, ...)
}
\arguments{
\item{object}{object of class "lrtest_msel"}

\item{...}{further arguments (currently ignored)}
}
\value{
Returns an object of class 'summary.lrtest_msel'.
}
\description{
Provides summary for an object of class 'lrtest_msel'.
}
\details{
This function just changes the class of the 'lrtest_msel'
object to 'summary.lrtest_msel'.
}
