% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{summary.test_msel}
\alias{summary.test_msel}
\title{Summary for an Object of Class delta_method}
\usage{
\method{summary}{test_msel}(object, ..., is_legend = TRUE)
}
\arguments{
\item{object}{object of class 'delta_method'}

\item{...}{further arguments (currently ignored)}

\item{is_legend}{a logical; if \code{TRUE} then additional information
is shown.}
}
\value{
Returns an object of class 'summary.delta_method'.
}
\description{
Provides summary for an object of class 'delta_method'.
}
