% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov_combine}
\alias{vcov_combine}
\title{Combine the asymptotic covariance matrices}
\usage{
vcov_combine(object, type = "cov")
}
\arguments{
\item{object}{a list of objects of class \code{msel}.}

\item{type}{the same as the \code{type} argument of the 
\code{\link[switchSelection]{vcov.msel}} function.}
}
\description{
Combine the asymptotic covariance matrices
}
