% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.msel}
\alias{fitted.msel}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{msel}(object, ..., newdata = NULL)
}
\arguments{
\item{object}{object of class 'msel'.}

\item{...}{further arguments (currently ignored).}

\item{newdata}{an optional data frame in which to look for variables 
with which to predict. If omitted, the original data frame used. 
This data frame should contain values of dependent variables even if they 
are not actually needed for prediction (simply assign them with 0 values).}
}
\value{
Returns a numeric matrix. Its columns which names coincide with the
names of the ordinal and multinomial equations provide the index of the most 
probable category for each observation.
Columns which names coincide with the names of the continuous equations 
provide conditional expectations of the dependent variables in observable 
regimes for each observation.
}
\description{
Extracts fitted values from 'msel' object
}
