% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnprobit.R
\name{print.mnprobit}
\alias{print.mnprobit}
\title{Print for an Object of Class mnprobit}
\usage{
\method{print}{mnprobit}(x, ...)
}
\arguments{
\item{x}{object of class "mnprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns input argument \code{x}.
}
\description{
Prints summary for an object of class 'mnprobit'.
}
