% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoprobit.R
\name{print.summary.mvoprobit}
\alias{print.summary.mvoprobit}
\title{Print summary for an Object of Class mvoprobit}
\usage{
\method{print}{summary.mvoprobit}(x, ...)
}
\arguments{
\item{x}{object of class "mvoprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
The function returns input argument \code{x} changing
it's class to \code{lrtest}.
}
\description{
Prints summary for an object of class 'mvoprobit'.
}
