% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{struct_msel}
\alias{struct_msel}
\title{Structure of the Object of Class msel}
\usage{
struct_msel(x)
}
\arguments{
\item{x}{object of class 'msel'}
}
\value{
The function returns a numeric matrix which columns are 
\code{groups}, \code{groups2}, \code{groups3} correspondingly. It also has
additional (last) column with the number of observations associated with the
corresponding combinations of the groups.
}
\description{
Prints information on the structure of the model.
}
