% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnprobit.R
\name{summary.mnprobit}
\alias{summary.mnprobit}
\title{Summary for an Object of Class mnprobit}
\usage{
\method{summary}{mnprobit}(object, ...)
}
\arguments{
\item{object}{object of class "mnprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
Returns an object of class 'summary.mnprobit'.
}
\description{
Provides summary for an object of class 'mnprobit'.
}
\details{
This function just changes the class of the 'mnprobit'
object to 'summary.mnprobit'.
}
