% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoprobit.R
\name{summary.mvoprobit}
\alias{summary.mvoprobit}
\title{Summary for an Object of Class mvoprobit}
\usage{
\method{summary}{mvoprobit}(object, ...)
}
\arguments{
\item{object}{object of class "mvoprobit"}

\item{...}{further arguments (currently ignored)}
}
\value{
Returns an object of class 'summary.mvoprobit'.
}
\description{
Provides summary for an object of class 'mvoprobit'.
}
\details{
This function just changes the class of the 'mvoprobit'
object to 'summary.mvoprobit'.
}
