% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{update_msel}
\alias{update_msel}
\title{Update msel object with the new estimates}
\usage{
update_msel(object, par)
}
\arguments{
\item{object}{an object of class \code{'msel'}.}

\item{par}{a vector of parameters which substitutes \code{object$par} and
used to update the estimates i.e., \code{object$coef}, \code{object$cuts} and
others.}
}
\value{
This function returns an object \code{object} of class \code{'msel'}
in which \code{object$par} is substituted with \code{par}. Also, \code{par} 
is used to update the estimates i.e., \code{object$coef}, \code{object$cuts} 
and others.
}
\description{
This function updates parameters of the model estimated via 
\code{\link[switchSelection]{msel}} function.
}
\details{
It may be useful to apply this function to the bootstrap
estimates of \code{\link[switchSelection]{bootstrap_msel}}.
}
